Changing the returned type.

+++ src/Algorithms/BitManipulation/ConvertNumberToBinary.cs
@@ -2,9 +2,9 @@
 {
     public class ConvertNumberToBinary
     {
-        public int[] FirstTry(sbyte number)
+        public byte[] FirstTry(sbyte number)
         {
-            int[] result = new int[8];
+            byte[] result = new byte[8];
             int index = result.Length - 1;
 
             while (number > 0)
@@ -12,7 +12,7 @@
                 int mod = number % 2;
                 number /= 2;
 
-                result[index--] = mod;
+                result[index--] = (byte)mod;
             }
 
             return result;
